@@ -13,6 +13,7 @@
     "cross-fetch": "^3.1.4",
     "fetcher-ts": "^1.1.1",
     "fp-ts": "^2.10.5",
+    "i18next": "^20.4.0",
     "io-ts": "^2.2.16",
     "mobx": "6.3.2",
     "mobx-react": "7.2.0",
@@ -20,6 +21,7 @@
     "react": "17.0.2",
     "react-bootstrap": "1.6.0",
     "react-dom": "17.0.2",
+    "react-i18next": "^11.11.4",
     "react-router-dom": "^5.2.0",
     "react-select": "^4.3.0",
     "reflect-metadata": "^0.1.13",
@@ -43,7 +45,7 @@
     "mobx-devtools-mst": "0.9.26",
     "style-loader": "2.0.0",
     "ts-loader": "9.1.1",
-    "typescript": "4.1.2",
+    "typescript": "4.4.4",
     "webpack": "5.36.2",
     "webpack-cli": "4.6.0",
     "webpack-merge": "5.7.3"
