@@ -1,29 +1,34 @@
 package org.vstu.compprehension.controllers;
 
 
+import lombok.SneakyThrows;
+import lombok.extern.log4j.Log4j2;
 import lombok.var;
+import net.oauth.server.OAuthServlet;
+import org.apache.commons.codec.net.URLCodec;
+import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.math.NumberUtils;
-import org.vstu.compprehension.Service.ExerciseService;
+import org.apache.jena.shared.NotFoundException;
 import org.vstu.compprehension.Service.UserService;
 import org.vstu.compprehension.dto.*;
-import org.vstu.compprehension.models.repository.ExerciseAttemptRepository;
 import lombok.val;
 import org.imsglobal.lti.launch.LtiOauthVerifier;
-import org.imsglobal.lti.launch.LtiVerificationResult;
-import org.imsglobal.lti.launch.LtiVerifier;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.vstu.compprehension.models.repository.ExerciseRepository;
 import org.vstu.compprehension.utils.Mapper;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("lti")
+@Log4j2
 public class LtiExerciseController extends BasicExerciseController {
     @Value("${config.property.lti_launch_secret}")
     private String ltiLaunchSecret;
@@ -31,13 +36,31 @@ public class LtiExerciseController extends BasicExerciseController {
     @Autowired
     private UserService userService;
 
-    @RequestMapping(value = {"/pages/exercise" }, method = { RequestMethod.POST })
-    public String ltiLaunch(Model model, HttpServletRequest request, @RequestParam Map<String, String> params) throws Exception {
-        LtiVerifier ltiVerifier = new LtiOauthVerifier();
-        String secret = this.ltiLaunchSecret;
-        LtiVerificationResult ltiResult = ltiVerifier.verify(request, secret);
+    @Autowired
+    private ExerciseRepository exerciseRepository;
+
+    @SneakyThrows
+    @RequestMapping(value = {"/pages/exercise" }, method=RequestMethod.POST)
+    public String ltiLaunch(Model model, HttpServletRequest request, @RequestParam Map<String, String> requestParams) {
+        // read LTI data from both request body and request params
+        // we can't just extract full form data explicitly through `@RequestParam` or something
+        // because we've already cached request and broken formdata->requestparams implicit conversion
+        val decodeCodec = new URLCodec();
+        val rawBody = IOUtils.toString(request.getInputStream(), request.getCharacterEncoding());
+        val params = Arrays.stream(rawBody.split("&"))
+            .map(x -> x.split("="))
+            .collect(Collectors.toMap(x -> x[0], x -> { try { return decodeCodec.decode(x[1]); } catch (Exception e) { return ""; }}));
+        params.putAll(requestParams);
+
+        // ensure LTI session validity
+        val ltiVerifier = new LtiOauthVerifier();
+        val secret = this.ltiLaunchSecret;
+        val ltiPreparedUrl = OAuthServlet.getMessage(request, null).URL; // special LTI url for correct own `secret` generation
+        val ltiResult = ltiVerifier.verifyParameters(params, ltiPreparedUrl, request.getMethod(), secret);
         if (!ltiResult.getSuccess()) {
-            throw new Exception("Invalid LTI session");
+            model.addAttribute("exerciseLaunchError", "Invalid LTI session. " + ltiResult.getError());
+            log.error("Invalid LTI session. " + ltiResult.getMessage());
+            return "index";
         }
 
         var session = request.getSession();
@@ -47,13 +70,25 @@ public class LtiExerciseController extends BasicExerciseController {
         }
         session.setAttribute("ltiSessionInfo", params);
 
-        val exerciseIdS = params.getOrDefault("custom_exerciseId", "-1");
-        val exerciseId = NumberUtils.toLong(exerciseIdS, -1L);
-        if (exerciseId == -1) {
-            throw new Exception("Param 'custom_exerciseId' is required");
+        val exerciseId = List.of("custom_exerciseId", "exerciseId").stream()
+                .map(prop -> params.getOrDefault(prop, "-1"))
+                .map(vl -> NumberUtils.toLong(vl, -1L))
+                .filter(v -> v != -1L)
+                .findFirst()
+                .orElse(-1L);
+        if (exerciseId == -1L) {
+            log.error("Param 'custom_exerciseId' or 'exerciseId' is required");
+            model.addAttribute("exerciseLaunchError", "Param 'custom_exerciseId' or 'exerciseId' is required");
+            return "index";
         }
 
-        return super.launch(exerciseId, request);
+        return super.launch(model, exerciseId, request);
+    }
+
+    @Override
+    public String launch(Model model, Long exerciseId, HttpServletRequest request) {
+        log.error("No LTI context found");
+        return super.launch(model, exerciseId, request);
     }
 
     @Override
@@ -71,6 +106,7 @@ public class LtiExerciseController extends BasicExerciseController {
         val userEntity = userService.createOrUpdateFromLti(params);
         val userEntityDto = Mapper.toDto(userEntity);
         session.setAttribute("currentUserInfo", userEntityDto);
+        session.setAttribute("currentUserId", userEntityDto.getId());
         return userEntityDto;
     }
 
@@ -87,11 +123,13 @@ public class LtiExerciseController extends BasicExerciseController {
             throw new Exception("Couldn't get session info");
         }
         val exerciseId = (Long)session.getAttribute("exerciseId");
+        val exercise = exerciseRepository.findById(exerciseId)
+                .orElseThrow(() -> new NotFoundException("exercise"));
         val user = getCurrentUser(request);
         val language = ltiParams.getOrDefault("launch_presentation_locale", "EN").toUpperCase();
         val sessionInfo = SessionInfoDto.builder()
                 .sessionId(session.getId())
-                .exerciseId(exerciseId)
+                .exercise(new ExerciseInfoDto(exerciseId, exercise.getOptions()))
                 .user(user)
                 .language(language)
                 .build();
