@@ -2,24 +2,28 @@ package org.vstu.compprehension.dto.feedback;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.jetbrains.annotations.NotNull;
-import org.vstu.compprehension.models.entities.QuestionOptions.MatchingQuestionOptionsEntity;
+import org.jetbrains.annotations.Nullable;
+import org.vstu.compprehension.dto.AnswerDto;
+import org.vstu.compprehension.models.entities.EnumData.Decision;
 
 @Data
 @SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
 public class FeedbackDto {
+    @JsonProperty("isCorrect")
+    private boolean isCorrect;
     private Float grade;
     private Integer correctSteps;
     private Integer stepsLeft;
     private Integer stepsWithErrors;
-    private Long[][] correctAnswers;
+    private AnswerDto[] correctAnswers;
     private Message[] messages;
+    private Decision strategyDecision;
 
     public enum MessageType {
         @JsonProperty("ERROR")
@@ -33,13 +37,16 @@ public class FeedbackDto {
     public static class Message {
         @NotNull private MessageType type;
         @NotNull private String message;
-        @NotNull private String[] violationLaws;
+        @Nullable private FeedbackViolationLawDto violationLaw;
 
-        public static Message Success(@NotNull String message, @NotNull String[] violationLaws) {
-            return new Message(MessageType.SUCCESS, message, violationLaws);
+        public static Message Success(@NotNull String message) {
+            return new Message(MessageType.SUCCESS, message, null);
         }
-        public static Message Error(@NotNull String message, @NotNull String[] violationLaws) {
-            return new Message(MessageType.ERROR, message, violationLaws);
+        public static Message Success(@NotNull String message, @Nullable FeedbackViolationLawDto violationLaw) {
+            return new Message(MessageType.SUCCESS, message, violationLaw);
+        }
+        public static Message Error(@NotNull String message, @Nullable FeedbackViolationLawDto violationLaw) {
+            return new Message(MessageType.ERROR, message, violationLaw);
         }
     }
 }
