@@ -1,8 +1,12 @@
 package org.vstu.compprehension.models.businesslogic.backend;
 
 
+import lombok.extern.log4j.Log4j2;
 import org.apache.jena.vocabulary.*;
+import org.springframework.web.context.annotation.RequestScope;
+import org.springframework.web.context.annotation.SessionScope;
 import org.vstu.compprehension.models.businesslogic.Law;
+import org.vstu.compprehension.models.businesslogic.domains.helpers.FactsGraph;
 import org.vstu.compprehension.models.entities.BackendFactEntity;
 import org.vstu.compprehension.models.businesslogic.LawFormulation;
 import org.apache.jena.datatypes.RDFDatatype;
@@ -13,23 +17,21 @@ import org.apache.jena.reasoner.rulesys.Rule;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.util.PrintUtil;
-import org.apache.jena.util.iterator.ExtendedIterator;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
-import java.io.ByteArrayInputStream;
+import javax.inject.Singleton;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
-import java.io.StringWriter;
 import java.util.*;
 
 import static org.apache.jena.ontology.OntModelSpec.OWL_MEM;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 
 @Primary
-@Component
+@Component @RequestScope
+@Log4j2
 public class JenaBackend extends Backend {
 
     static String BACKEND_TYPE = "Jena";
@@ -157,8 +159,7 @@ public class JenaBackend extends Backend {
                 try {
                     domainRules.add(Rule.parseRule(lawFormulation.getFormulation()));
                 } catch (Rule.ParserException e) {
-                    System.out.println("Following error in rule: " + lawFormulation.getFormulation());
-                    e.printStackTrace();
+                    log.error("Following error in rule: " + lawFormulation.getFormulation(), e);
                 }
             } else if ("can_covert" != null) {
 
@@ -175,8 +176,9 @@ public class JenaBackend extends Backend {
         String obj = fact.getObject();
         String objType = fact.getObjectType();
         assert prop != null;
+        assert objType != null;
 
-//        // debug
+        /// debug
 //        System.out.println("addStatementFact( subj: " + subj + ", prop: " + prop + ", obj: " + obj + " )");
 
         // TODO: save this info somehow?
@@ -262,7 +264,7 @@ public class JenaBackend extends Backend {
 
         long estimatedTime = System.nanoTime() - startTime;
         // print time report. TODO: remove the print
-        System.out.println("Time Jena spent on reasoning: " + String.valueOf((float) (estimatedTime / 1000) / 1000 / 1000) + " seconds.");
+        log.info("Time Jena spent on reasoning: " + String.format("%.5f", (float)estimatedTime / 1000 / 1000 / 1000) + " seconds.");
 
         // use the inferred results (inf) ...
         model.add( inf );
@@ -276,6 +278,7 @@ public class JenaBackend extends Backend {
     /** Expand simple name as local, prefixed name as special */
     private String termToUri(String s) {
         String uri;
+        assertNotNull(s, "termToUri(null) !");
         if (s.startsWith("http://")) {
             uri = s;
         }
@@ -297,7 +300,7 @@ public class JenaBackend extends Backend {
     /** Returns local name or prefixed special name */
     private String uriToTerm(String uri) {
         if (uri == null) {
-            System.out.println("uriToTerm(): Encountered NULL uri! Defaulting to '[]'");
+            log.debug("uriToTerm(): Encountered NULL uri! Defaulting to '[]'");
             return "[]";
         }
 
@@ -312,7 +315,7 @@ public class JenaBackend extends Backend {
         }
 
         ///
-        // System.out.println("uriToTerm: " + uri + " -> " + s);
+        // System.out.println("uriToTerm: " + uri + "\n -> " + s);
         ///
         return s;
     }
@@ -344,7 +347,12 @@ public class JenaBackend extends Backend {
 
             if (isObjectProp) {
                 Resource resource = objNode.asResource();
-                obj = uriToTerm(resource.getURI());
+                String URI = resource.getURI();
+                if (URI != null) {
+                    obj = uriToTerm(URI);
+                } else { /* resource is bnode*/
+                    obj = resource.toString();
+                }
 
                 if (resource.hasProperty(RDF.type, OWL.Class))
                     objType = "owl:Class";
@@ -361,9 +369,17 @@ public class JenaBackend extends Backend {
                 obj = objNode.asLiteral().getLexicalForm();
             }
 
+            String subjURI = subjResource.getURI();
+            String subj;
+            if (subjURI != null) {
+                subj = uriToTerm(subjURI);
+            } else { /* subj is bnode*/
+                subj = subjResource.toString();
+            }
+
             facts.add(new BackendFactEntity(
                     subjType,
-                    uriToTerm(subjResource.getURI()),
+                    subj,
                     propName,
                     objType,
                     obj
@@ -384,13 +400,13 @@ public class JenaBackend extends Backend {
             String uri = termToUri(verb);
             OntResource resource = model.getOntResource(uri);
             if (resource == null) {
-                System.out.println("JenaBack.getFacts() WARNING: Cannot find resource for verb: " + verb);
+                log.warn("JenaBack.getFacts() WARNING: Cannot find resource for verb: " + verb);
                 continue;
             }
             try {
                 p = resource.asProperty();
             } catch (ConversionException exception) {
-                System.out.println("JenaBack.getFacts() WARNING: Cannot find property for verb: " + verb);
+                log.warn("JenaBack.getFacts() WARNING: Cannot find property for verb: " + verb);
                 continue;
             }
 
@@ -410,10 +426,9 @@ public class JenaBackend extends Backend {
             try {
                 out = new FileOutputStream(out_rdf_path);
                 RDFDataMgr.write(out, model, Lang.NTRIPLES);  // Lang.NTRIPLES  or  Lang.RDFXML
-                System.out.println("Debug written: " + out_rdf_path);
+                log.debug("Debug written: " + out_rdf_path + ". N of of triples: " + model.size());
             } catch (FileNotFoundException e) {
-                e.printStackTrace();
-                System.out.println("Cannot write to file: " + out_rdf_path);
+                log.error("Cannot write to file: " + out_rdf_path, e);
             }
         }
     }
@@ -433,7 +448,7 @@ public class JenaBackend extends Backend {
 
         debug_dump_model("solved");
 
-        return getFacts(solutionVerbs);
+        return new FactsGraph(getFacts(solutionVerbs)).removeDuplicates().getFacts();
     }
 
     public void addFacts(List<BackendFactEntity> facts) {
@@ -460,7 +475,7 @@ public class JenaBackend extends Backend {
 
         debug_dump_model("judged");
 
-        return getFacts(violationVerbs);
+        return new FactsGraph(getFacts(violationVerbs)).removeDuplicates().getFacts();
     }
 
 
