@@ -1,9 +1,7 @@
 package org.vstu.compprehension.models.entities.EnumData;
 
 public enum AttemptStatus {
-    COMPLETE,
-    INCOMPLETE;
-
-    AttemptStatus() {
-    }
+    COMPLETED_BY_USER,
+    COMPLETED_BY_SYSTEM,
+    INCOMPLETE,
 }
