@@ -1,6 +1,9 @@
 package org.vstu.compprehension.models.entities;
 
+import com.vladmihalcea.hibernate.type.json.JsonStringType;
 import lombok.ToString;
+import org.hibernate.annotations.Type;
+import org.hibernate.annotations.TypeDef;
 import org.vstu.compprehension.models.businesslogic.Tag;
 import org.vstu.compprehension.models.entities.EnumData.Complexity;
 import org.vstu.compprehension.models.entities.EnumData.ExerciseType;
@@ -21,6 +24,7 @@ import java.util.stream.Collectors;
 @Data
 @NoArgsConstructor
 @Table(name = "Exercise")
+@TypeDef(name = "json", typeClass = JsonStringType.class)
 public class ExerciseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -45,6 +49,10 @@ public class ExerciseEntity {
     @Column(name = "tags")
     private String tags;
 
+    @Type(type = "json")
+    @Column(name = "options_json", columnDefinition = "json", nullable = false)
+    private ExerciseOptionsEntity options;
+
     public List<Tag> getTags() {
         return Arrays.stream(tags.split(","))
                 .map(i -> new Tag(i))
@@ -63,12 +71,6 @@ public class ExerciseEntity {
     @Enumerated(EnumType.ORDINAL)
     private Language language;
 
-
-    @ToString.Exclude
-    @ManyToOne
-    @JoinColumn(name = "course_id", nullable = false)
-    private CourseEntity course;
-
     @ManyToOne
     @JoinColumn(name = "backend_id", nullable = false)
     private BackendEntity backend;
@@ -83,13 +85,7 @@ public class ExerciseEntity {
     private List<AdditionalFieldEntity> additionalFields;
 
     @OneToMany(mappedBy = "exercise", fetch = FetchType.LAZY)
-    private List<ExerciseDisplayingFeedbackTypeEntity> exerciseDisplayingFeedbackTypes;
-
-    @OneToMany(mappedBy = "exercise", fetch = FetchType.LAZY)
     private List<ExerciseQuestionTypeEntity> exerciseQuestionTypes;
-
-    @OneToMany(mappedBy = "exercise", fetch = FetchType.LAZY)
-    private List<UserActionExerciseEntity> userActionExercises;
 
     @OneToMany(mappedBy = "exercise", fetch = FetchType.LAZY)
     private List<ExerciseLawsEntity> exerciseLaws;
