@@ -11,6 +11,7 @@ import org.hibernate.annotations.NotFoundAction;
 
 import javax.persistence.*;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -35,6 +36,9 @@ public class InteractionEntity {
     @Enumerated(EnumType.STRING)
     private InteractionType interactionType;
 
+    @Column(nullable = true, columnDefinition="DATETIME(6)")
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date date;
 
     @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     @JoinColumn(name = "feedback_id", referencedColumnName = "id")
@@ -50,9 +54,12 @@ public class InteractionEntity {
     private List<CorrectLawEntity> correctLaw;
 
     @ToString.Exclude
-    @OneToMany(mappedBy = "interaction", fetch = FetchType.LAZY)
+    @OneToMany(mappedBy = "interaction", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     private List<ResponseEntity> responses;
 
+    @ToString.Exclude
+    @OneToMany(mappedBy = "createdByInteraction", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    private List<ResponseEntity> newResponses;
 
     @ToString.Exclude
     @ManyToOne
@@ -64,7 +71,8 @@ public class InteractionEntity {
             QuestionEntity question,
             List<ViolationEntity> violations,
             List<String> correctlyAppliedLaws,
-            List<ResponseEntity> responses){
+            List<ResponseEntity> allResponses,
+            List<ResponseEntity> newResponses){
         this.setQuestion(question);
         this.setInteractionType(type);
         this.setFeedback(new FeedbackEntity());
@@ -74,9 +82,14 @@ public class InteractionEntity {
             m.setInteraction(this);
         }
 
-        this.setResponses(new ArrayList<>(responses));
+        this.setResponses(new ArrayList<>(allResponses));
         for(val r : this.getResponses()) {
             r.setInteraction(this);
+        }
+
+        this.setNewResponses(new ArrayList<>(newResponses));
+        for(val r : this.getNewResponses()) {
+            r.setCreatedByInteraction(this);
         }
 
         if(correctlyAppliedLaws == null){
@@ -92,5 +105,6 @@ public class InteractionEntity {
                     .collect(Collectors.toList());
             this.setCorrectLaw(correctLaw);
         }
+        this.date = new Date();
     }
 }
