@@ -1,7 +1,6 @@
 package org.vstu.compprehension.utils;
 
 import lombok.val;
-import org.apache.commons.lang3.tuple.Pair;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.vstu.compprehension.dto.*;
@@ -13,15 +12,15 @@ import org.vstu.compprehension.dto.question.QuestionDto;
 import org.vstu.compprehension.models.businesslogic.Question;
 import org.vstu.compprehension.models.businesslogic.domains.Domain;
 import org.vstu.compprehension.models.entities.*;
+import org.vstu.compprehension.models.entities.EnumData.Decision;
+import org.vstu.compprehension.models.entities.EnumData.InteractionType;
 import org.vstu.compprehension.models.entities.EnumData.QuestionType;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import java.util.stream.LongStream;
 import java.util.stream.Stream;
 
 public class Mapper {
@@ -35,18 +34,30 @@ public class Mapper {
                 .displayName(displayName)
                 .email(user.getEmail())
                 .roles(user.getRoles())
+                .preferredLanguage(user.getPreferred_language())
                 .build();
     }
 
+    public static @NotNull AnswerDto toDto(@NotNull ResponseEntity response) {
+        return AnswerDto.builder()
+                .isCreatedByUser(response.getCreatedByInteraction().getInteractionType() == InteractionType.SEND_RESPONSE)
+                .createdByInteraction(response.getCreatedByInteraction().getId())
+                .answer(new Long[] { (long)response.getLeftAnswerObject().getAnswerId(), (long)response.getRightAnswerObject().getAnswerId() })
+                .build();
+    }
+
+    /*
     public static @NotNull CorrectAnswerDto toDto(@NotNull Domain.CorrectAnswer correctAnswer) {
         val frontAnswers = Optional.ofNullable(correctAnswer.answers).stream()
                 .flatMap(Collection::stream)
-                .map(pair -> List.of((long)pair.getLeft().getAnswerId(), (long)pair.getRight().getAnswerId()).toArray(new Long[2])).toArray(Long[][]::new);
+                .map(answr -> new AnswerDto((long)answr.getLeft().getAnswerId(), (long)answr.getRight().getAnswerId(), answr.isCreatedByUser()))
+                .toArray(AnswerDto[]::new);
         return CorrectAnswerDto.builder()
                 .explanation(correctAnswer.explanation.getText())
                 .answers(frontAnswers)
                 .build();
     }
+    */
 
     public static @NotNull QuestionDto toDto(@NotNull Question questionObject) {
         val question = questionObject.getQuestionData();
@@ -70,12 +81,11 @@ public class Mapper {
         val responses = lastCorrectInteraction
                 .flatMap(i -> Optional.ofNullable(i.getResponses())).stream()
                 .flatMap(Collection::stream)
-                .map(r -> Pair.of(r.getLeftAnswerObject(), r.getRightAnswerObject()))
-                .map(pair -> List.of((long)pair.getLeft().getAnswerId(), (long)pair.getRight().getAnswerId()).toArray(new Long[2]))
-                .toArray(Long[][]::new);
+                .map(Mapper::toDto)
+                .toArray(AnswerDto[]::new);
 
         val feedback = lastInteraction
-                .map(i -> Mapper.toFeedbackDto(questionObject, i, null, correctInteractionsCount, interactionsWithErrorsCount, null))
+                .map(i -> Mapper.toFeedbackDto(questionObject, null, correctInteractionsCount, interactionsWithErrorsCount, i.getFeedback().getGrade(), i.getFeedback().getInteractionsLeft(), null, i.getViolations().size() == 0, null))
                 .orElse(null);
 
         val answers = question.getAnswerObjects() != null ? question.getAnswerObjects() : new ArrayList<AnswerObjectEntity>(0);
@@ -84,7 +94,7 @@ public class Mapper {
                 .toArray(QuestionAnswerDto[]::new);
         switch (question.getQuestionType()) {
             case ORDER:
-                val domain = DomainAdapter.getDomain(question.getDomainEntity().getName());
+                val domain = DomainAdapter.getDomain(question.getDomainEntity().getClassPath());
                 val trace = Optional.ofNullable(domain.getFullSolutionTrace(questionObject)).stream()
                         .flatMap(Collection::stream)
                         .map(HyperText::getText)
@@ -153,14 +163,17 @@ public class Mapper {
 
     public static @NotNull FeedbackDto toFeedbackDto(
             @NotNull Question question,
-            @NotNull InteractionEntity interaction,
             @Nullable FeedbackDto.Message[] messages,
             @Nullable Integer correctSteps,
             @Nullable Integer stepsWithErrors,
-            @Nullable Long[][] correctAnswers
+            @Nullable Float grade,
+            @Nullable Integer interactionsLeft,
+            @Nullable AnswerDto[] correctAnswers,
+            boolean isCorrect,
+            @Nullable Decision strategyDecision
     ) {
-        if (interaction.getQuestion().getQuestionType() == QuestionType.ORDER) {
-            val domain = DomainAdapter.getDomain(question.getQuestionData().getDomainEntity().getName());
+        if (question.getQuestionData().getQuestionType() == QuestionType.ORDER) {
+            val domain = DomainAdapter.getDomain(question.getQuestionData().getDomainEntity().getClassPath());
             val trace = Optional.ofNullable(domain.getFullSolutionTrace(question)).stream()
                     .flatMap(Collection::stream)
                     .map(HyperText::getText)
@@ -168,20 +181,24 @@ public class Mapper {
             return OrderQuestionFeedbackDto.builder()
                     .correctSteps(correctSteps)
                     .stepsWithErrors(stepsWithErrors)
-                    .grade(interaction.getFeedback().getGrade())
+                    .grade(grade)
                     .messages(messages)
                     .correctAnswers(correctAnswers)
-                    .stepsLeft(interaction.getFeedback().getInteractionsLeft())
+                    .stepsLeft(interactionsLeft)
+                    .strategyDecision(strategyDecision)
                     .trace(trace)
+                    .isCorrect(isCorrect)
                     .build();
         }
         return FeedbackDto.builder()
                 .correctSteps(correctSteps)
                 .stepsWithErrors(stepsWithErrors)
-                .grade(interaction.getFeedback().getGrade())
+                .grade(grade)
                 .messages(messages)
                 .correctAnswers(correctAnswers)
-                .stepsLeft(interaction.getFeedback().getInteractionsLeft())
+                .stepsLeft(interactionsLeft)
+                .strategyDecision(strategyDecision)
+                .isCorrect(isCorrect)
                 .build();
     }
 }
