@@ -1,18 +1,21 @@
+import { observer } from "mobx-react";
 import React from "react";
 import { useEffect } from "react";
+import { Answer } from "../../../types/answer";
 import { OrderQuestionFeedback } from "../../../types/feedback";
 import { OrderQuestion } from "../../../types/question";
-import { notNullOrUndefinded } from "../../../utils/helpers";
+import { notNulAndUndefinded } from "../../../utils/helpers";
 import { Optional } from "../optional";
 
 type OrderQuestionComponentProps = {
     question: OrderQuestion,
-    feedback?: OrderQuestionFeedback,
-    answers: [number, number][],
-    getAnswers: () => [number, number][],
-    onChanged: (newAnswers: [number, number][]) => void,
+    getFeedback: () => OrderQuestionFeedback | undefined, 
+    answers: Answer[],
+    getAnswers: () => Answer[],
+    onChanged: (newAnswers: Answer[]) => void,
 }
-export const OrderQuestionComponent = ({ question, getAnswers, onChanged, feedback }: OrderQuestionComponentProps) => {
+export const OrderQuestionComponent = observer((props: OrderQuestionComponentProps) => {
+    const { question, getAnswers, onChanged, getFeedback } = props; 
     if (!question.options.requireContext) {
         return null;
     }
@@ -27,7 +30,7 @@ export const OrderQuestionComponent = ({ question, getAnswers, onChanged, feedba
         document.querySelectorAll(`[id^="question_${question.questionId}_answer_"]`).forEach(e => {
             const idStr = e.id?.split(`question_${question.questionId}_answer_`)[1] ?? ""; 
             const id = +idStr;
-            e.addEventListener('click', () => onChanged([...getAnswers(), [id, id]]));
+            e.addEventListener('click', () => onChanged([...getAnswers(), { answer: [id, id], isСreatedByUser: true }]));
         })
 
         // show elements positions
@@ -44,13 +47,34 @@ export const OrderQuestionComponent = ({ question, getAnswers, onChanged, feedba
             const pos = e.getAttribute("data-comp-ph-pos");
             e.innerHTML = originalText.querySelector(`#${e.id}`)?.innerHTML + (pos ? `<span class="comp-ph-expr-top-hint">${pos}</span>` : '')
             e.classList.remove('disabled');
+            e.classList.remove('comp-ph-question-answer--last-selected-by-system');
         });
 
+        // calc last generated by system answers
+        const isLastAsnwerCorrect = getFeedback()?.isCorrect ?? true;
+        const lastGeneratedAnswers = !isLastAsnwerCorrect
+            ? []
+            : [...getAnswers()].reverse().reduce<[number, Answer[]]>(((acc, answer) => {
+                const [prevInteractionId, result] = acc;
+                if (prevInteractionId === -1)
+                    return acc;
+                if (answer.isСreatedByUser || prevInteractionId !== 0 && prevInteractionId !== answer.createdByInteraction)
+                    return [-1, result];
+                result.push(answer);
+                return [answer.createdByInteraction || -1, result];
+            }), [0, []])[1].reverse();
+
         // apply history changes    
-        getAnswers().forEach(([h], idx) => {
+        getAnswers().forEach((answer, idx, answers) => {
+            const { answer:asnwerPair } = answer;
+            const [h] = asnwerPair;
             const answr = document.querySelector(`#question_${question.questionId}_answer_${h}`);
             if (!answr) {
                 return 0;
+            }
+
+            if (lastGeneratedAnswers.includes(answer)) {
+                answr.classList.add('comp-ph-question-answer--last-selected-by-system');
             }
 
             // add pos hint        
@@ -61,18 +85,18 @@ export const OrderQuestionComponent = ({ question, getAnswers, onChanged, feedba
                 answr.innerHTML = answerHtml.join(orderNumberOptions.delimiter);
             }  
             // disable if needed
-            if (options.multipleSelectionEnabled) {            
-                answr.classList.add('disabled');            
+            if (!options.multipleSelectionEnabled) {            
+                answr.classList.add('disabled');
             }
         });
     }, [question.questionId, getAnswers().length])
     
-    const trace = feedback?.trace ?? (getAnswers().length === 0 ? question.initialTrace : null);
-    const isTraceVisible = options.showTrace && notNullOrUndefinded(trace) && trace.length > 0;
+    const trace = getFeedback()?.trace ?? (getAnswers().length === 0 ? question.initialTrace : null);
+    const isTraceVisible = options.showTrace && notNulAndUndefinded(trace) && trace.length > 0;
 
     return (
         <div>
-            <div dangerouslySetInnerHTML={{ __html: question.text }} />
+            <div className="comp-ph-question-text" dangerouslySetInnerHTML={{ __html: question.text }} />
             <Optional isVisible={isTraceVisible}>
                 <p>
                     <table className="comp-ph-trace">
@@ -84,4 +108,4 @@ export const OrderQuestionComponent = ({ question, getAnswers, onChanged, feedba
             </Optional>
         </div>
     );
-}
+})
