@@ -1,4 +1,7 @@
+import { absurd } from "fp-ts/lib/function";
+import { observer } from "mobx-react";
 import React from "react";
+import { Answer } from "../../../types/answer";
 import { Feedback } from "../../../types/feedback";
 import { Question } from "../../../types/question"
 import { MatchingQuestionComponent } from "./matching-question";
@@ -8,23 +11,46 @@ import { SingleChoiceQuestionComponent } from "./single-choice-question";
 
 type QuestionComponentProps = {
     question: Question,
-    feedback?: Feedback, 
-    answers: [number, number][],
-    getAnswers: () => [number, number][],
-    onChanged: (x: [number, number][]) => void,
+    getFeedback: () => Feedback | undefined,
+    isFeedbackLoading: boolean,
+    isQuestionFreezed: boolean,
+    answers: Answer[],
+    getAnswers: () => Answer[],
+    onChanged: (x: Answer[]) => void,
 }
 
-export const QuestionComponent = (props: QuestionComponentProps) => {
-    const { question, answers, onChanged, getAnswers, feedback } = props;
+export const QuestionComponent = observer((props: QuestionComponentProps) => {
+    const { question, answers, onChanged, getAnswers, getFeedback, isFeedbackLoading, isQuestionFreezed } = props;
+    let questonComponent: JSX.Element;
     switch(question.type) {
         case 'MATCHING':                
-            return <MatchingQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers}/>;
+            questonComponent = <MatchingQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers}/>;
+            break;
         case 'MULTI_CHOICE':
-            return <MultiChoiceQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers}/>;
+            questonComponent = <MultiChoiceQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers}/>;
+            break;
         case 'SINGLE_CHOICE':
-            return <SingleChoiceQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers}/>;
+            questonComponent = <SingleChoiceQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers}/>;
+            break;
         case 'ORDER':
-            return <OrderQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers} feedback={feedback}/>;
+            questonComponent = <OrderQuestionComponent question={question} onChanged={onChanged} answers={answers} getAnswers={getAnswers} getFeedback={getFeedback}/>;
+            break;
+        default:
+            // compile-time checking whether the question has `never` type 
+            // to ensure that all case branches have been processed
+            return absurd<JSX.Element>(question);
     }
-    return (<div>Not implemented</div>);
-}
+
+    const wrapperClassName = [
+        "comp-ph-question-wrapper",
+        getFeedback()?.stepsLeft === 0 && "comp-ph-question-wrapper--finished" || "",
+        isFeedbackLoading && "comp-ph-question-wrapper--loading-feedback" || "",
+        isQuestionFreezed && "comp-ph-question-wrapper--freezed" || "",
+    ].join(' ');
+
+    return (
+        <div className={wrapperClassName}>
+            {questonComponent}
+        </div>
+    );
+})
