@@ -6,7 +6,9 @@ import { Interaction } from "../../types/interaction";
 import { Question, TOptionalQuestion, TQuestion } from "../../types/question";
 import { SessionInfo, TSessionInfo } from "../../types/session-info";
 import { SupplementaryQuestionRequest } from "../../types/supplementary-question-request";
-import { ajaxGet, ajaxPost, PromiseEither, RequestError } from "../../utils/ajax";
+import { ajaxGet, ajaxPost, PromiseEither } from "../../utils/ajax";
+import * as io from 'io-ts'
+import { RequestError } from "../../types/request-error";
 
 
 export interface IExerciseController {
@@ -19,15 +21,16 @@ export interface IExerciseController {
     getExistingExerciseAttempt(exerciseId: number): PromiseEither<RequestError, ExerciseAttempt | null | undefined | ''>;
     createExerciseAttempt(exerciseId: number): PromiseEither<RequestError, ExerciseAttempt>;
     getExerciseStatistics(exerciseId: number): PromiseEither<RequestError, ExerciseStatisticsItem[]>;
+    getExercises(): PromiseEither<RequestError, number[]>
 }
 
 @injectable()
 export class ExerciseController implements IExerciseController {
-    static endpointPath: string = ExerciseController.initEndpointPath();
-    private static initEndpointPath() {
+    static endpointPath: `/` | `/${string}/` = ExerciseController.initEndpointPath();
+    private static initEndpointPath(): `/` | `/${string}/` {
         const matches = /^\/(?:.+\/(?<!\/pages\/))?/.exec(window.location.pathname)
         if (matches) {
-            return matches[0];
+            return <`/${string}/`>matches[0];
         }
         return "/";
     }
@@ -66,5 +69,9 @@ export class ExerciseController implements IExerciseController {
 
     getExerciseStatistics(exerciseId: number): PromiseEither<RequestError, ExerciseStatisticsItem[]> {
         return ajaxGet(`${ExerciseController.endpointPath}getExerciseStatistics?exerciseId=${exerciseId}`, TExerciseStatisticsItems)
+    }
+
+    getExercises(): PromiseEither<RequestError, number[]> {
+        return ajaxGet(`${ExerciseController.endpointPath}getExercises`, io.array(io.number));
     }
 }
