@@ -1,6 +1,6 @@
 import { observer } from "mobx-react"
 import React, { useEffect, useState } from "react"
-import { IExerciseController } from "../controllers/exercise/exercise-controller";
+import { ExerciseController, IExerciseController } from "../controllers/exercise/exercise-controller";
 import * as E from "fp-ts/lib/Either";
 import { ExerciseStatisticsItem } from "../types/exercise-statistics";
 import { LoadingWrapper } from "../components/common/loader";
@@ -18,7 +18,7 @@ export const Statistics = () => {
                 throw new Error("Invalid exerciseId url param");
             }
 
-            const controller = container.resolve<IExerciseController>("ExerciseController");
+            const controller = container.resolve<IExerciseController>(ExerciseController);
             setIsLoading(true);
             const statistics = await controller.getExerciseStatistics(+exerciseId);
             if (E.isRight(statistics)) {
