@@ -1,13 +1,14 @@
 
 import * as io from 'io-ts'
+import { Answer, TAnswer } from './answer';
 
 export type Interaction = {
     attemptId: number,
     questionId: number,
-    answers: [number, number][],    
+    answers: Answer[],    
 }
 export const TInteraction : io.Type<Interaction> = io.type({
     attemptId: io.number,
     questionId: io.number,
-    answers: io.array(io.tuple([io.number, io.number])),
+    answers: io.array(TAnswer),
 }, 'Interaction');
